+++ app/services/sheets_context/pdf/pathfinder2/template.rb
@@ -4,7 +4,7 @@ module SheetsContext
   module Pdf
     module Pathfinder2
       class Template < SheetsContext::Pdf::Template
-        # rubocop: disable Metrics/AbcSize, Layout/LineLength, Metrics/MethodLength, Metrics/PerceivedComplexity
+        # rubocop: disable Metrics/AbcSize, Layout/LineLength, Metrics/MethodLength, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
         def to_pdf(character:)
           super
 
@@ -71,7 +71,7 @@ module SheetsContext
 
           render
         end
-        # rubocop: enable Metrics/AbcSize, Layout/LineLength, Metrics/MethodLength, Metrics/PerceivedComplexity
+        # rubocop: enable Metrics/AbcSize, Layout/LineLength, Metrics/MethodLength, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
 
         private